Fix rune asset in `getPoolPriceValue` (review
suggestion) and move it into `poolHelpers`
