+++ src/renderer/components/wallet/assets/AssetsTableCollapsable.tsx
@@ -10,7 +10,7 @@ import * as FP from 'fp-ts/lib/function'
 import * as O from 'fp-ts/lib/Option'
 import { useIntl } from 'react-intl'
 
-import { getPoolPriceValue } from '../../../services/binance/utils'
+import { getPoolPriceValue } from '../../../helpers/poolHelper'
 import { WalletBalancesRD } from '../../../services/clients'
 import { PoolDetails } from '../../../services/midgard/types'
 import { ApiError, ChainBalance, ChainBalances } from '../../../services/wallet/types'
