+++ src/renderer/i18n/ru/wallet.ts
@@ -8,7 +8,7 @@ const wallet: WalletMessages = {
   'wallet.column.balance': 'Баланс',
   'wallet.column.value': 'Количество',
   'wallet.action.send': 'Отправить',
-  'wallet.action.upgrade': 'Upgrade - RU',
+  'wallet.action.upgrade': 'Перевести',
   'wallet.action.receive': 'Получить',
   'wallet.action.remove': 'Удалить кошелек',
   'wallet.action.unlock': 'Разблокировать',
